{"schema":"decodex/commit/1","summary":"Modularize capture host scroll minimap renderer","authority":"manual"}
